requirements.txt
langchain-core
langchain-openai
langgraph
python-dotenv
openai==1.55.3
streamlit
mysql-connector-python